+++ tests/src/Command/CoverageCheckCommandTest.php
@@ -70,6 +70,7 @@ final class CoverageCheckCommandTest extends TestCase
             'no_children'  => \dirname(__FILE__, 3) . '/fixtures/no_children.xml',
             'no_metrics'   => \dirname(__FILE__, 3) . '/fixtures/no_project_metrics.xml',
             'thisLibrary'  => \dirname(__FILE__, 3) . '/fixtures/self_clover.xml',
+            'openClover'   => \dirname(__FILE__, 3) . '/fixtures/open_clover.xml',
         ];
 
         $coverageCheckCommand = new CoverageCheckCommand(new CoverageCheck());
@@ -287,17 +288,23 @@ final class CoverageCheckCommandTest extends TestCase
 
         $eol = PHP_EOL;
 
-        $expected = '------------------------------------------------------------------- -------------------------- ---------- ' . $eol;
-        $expected .= '  File                                                                Elements (Covered/Total)   Coverage  ' . $eol;
-        $expected .= ' ------------------------------------------------------------------- -------------------------- ---------- ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\Application.php                    10/10                      100.00%   ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\Command\CoverageCheckCommand.php   77/77                      100.00%   ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\CoverageCheck.php                  63/63                      100.00%   ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\Style\CoverageCheckStyle.php       4/4                        100.00%   ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\Utils.php                          16/16                      100.00%   ' . $eol;
-        $expected .= ' ------------------------------------------------------------------- -------------------------- ---------- ' . $eol;
-        $expected .= '  Overall Totals                                                      170/170                    100.00%   ' . $eol;
-        $expected .= ' ------------------------------------------------------------------- -------------------------- ----------';
+        $expected = '------------------------------------------------------------------------ -------------------------- ---------- ' . $eol;
+        $expected .= '  File                                                                     Elements (Covered/Total)   Coverage  ' . $eol;
+        $expected .= ' ------------------------------------------------------------------------ -------------------------- ---------- ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Application.php                         12/12                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\CoverageCheck.php                       82/82                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Utils.php                               37/37                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Command\CoverageCheckCommand.php        79/79                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Exceptions\InvalidInputFileException.   2/2                        100.00%   ' . $eol;
+        $expected .= '  php                                                                                                           ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Exceptions\NotAValidCloverFileExcepti   2/2                        100.00%   ' . $eol;
+        $expected .= '  on.php                                                                                                        ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Exceptions\ThresholdOutOfBoundsExcept   2/2                        100.00%   ' . $eol;
+        $expected .= '  ion.php                                                                                                       ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Style\CoverageCheckStyle.php            14/14                      100.00%   ' . $eol;
+        $expected .= ' ------------------------------------------------------------------------ -------------------------- ---------- ' . $eol;
+        $expected .= '  Overall Totals                                                           230/230                    100.00%   ' . $eol;
+        $expected .= ' ------------------------------------------------------------------------ -------------------------- ----------';
 
         self::assertSame($expected, trim($this->applicationTester->getDisplay()));
         self::assertSame(Command::SUCCESS, $this->applicationTester->getStatusCode());
@@ -327,6 +334,49 @@ final class CoverageCheckCommandTest extends TestCase
 
         self::assertSame($expected, trim($this->applicationTester->getDisplay()));
         self::assertSame(Command::FAILURE, $this->applicationTester->getStatusCode());
+    }
+
+    public function testShowFilesTableOutputCloverAndOpenCloverTheSame(): void
+    {
+        $this->applicationTester->run([
+            'cloverfile'   => self::$fixtures['thisLibrary'],
+            'threshold'    => 90,
+            '--show-files' => true,
+        ]);
+
+        self::assertSame(self::$fixtures['thisLibrary'], $this->applicationTester->getInput()->getArgument('cloverfile'));
+        self::assertSame(90, $this->applicationTester->getInput()->getArgument('threshold'));
+
+        $eol = PHP_EOL;
+
+        $expected = '------------------------------------------------------------------------ -------------------------- ---------- ' . $eol;
+        $expected .= '  File                                                                     Elements (Covered/Total)   Coverage  ' . $eol;
+        $expected .= ' ------------------------------------------------------------------------ -------------------------- ---------- ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Application.php                         12/12                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\CoverageCheck.php                       82/82                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Utils.php                               37/37                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Command\CoverageCheckCommand.php        79/79                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Exceptions\InvalidInputFileException.   2/2                        100.00%   ' . $eol;
+        $expected .= '  php                                                                                                           ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Exceptions\NotAValidCloverFileExcepti   2/2                        100.00%   ' . $eol;
+        $expected .= '  on.php                                                                                                        ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Exceptions\ThresholdOutOfBoundsExcept   2/2                        100.00%   ' . $eol;
+        $expected .= '  ion.php                                                                                                       ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Style\CoverageCheckStyle.php            14/14                      100.00%   ' . $eol;
+        $expected .= ' ------------------------------------------------------------------------ -------------------------- ---------- ' . $eol;
+        $expected .= '  Overall Totals                                                           230/230                    100.00%   ' . $eol;
+        $expected .= ' ------------------------------------------------------------------------ -------------------------- ----------';
+
+        self::assertSame($expected, trim($this->applicationTester->getDisplay()));
+        self::assertSame(Command::SUCCESS, $this->applicationTester->getStatusCode());
+
+        $this->applicationTester->run([
+            'cloverfile'   => self::$fixtures['openClover'],
+            'threshold'    => 90,
+            '--show-files' => true,
+        ]);
+
+        self::assertSame($expected, trim($this->applicationTester->getDisplay()));
     }
 
     public function testShowFilesTableOutputEmpty(): void
@@ -361,17 +411,20 @@ final class CoverageCheckCommandTest extends TestCase
 
         $eol = PHP_EOL;
 
-        $expected = '------------------------------------------------------------------- -------------------------- ---------- ' . $eol;
-        $expected .= '  File                                                                Elements (Covered/Total)   Coverage  ' . $eol;
-        $expected .= ' ------------------------------------------------------------------- -------------------------- ---------- ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\Application.php                    10/10                      100.00%   ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\Command\CoverageCheckCommand.php   77/77                      100.00%   ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\CoverageCheck.php                  63/63                      100.00%   ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\Style\CoverageCheckStyle.php       4/4                        100.00%   ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\Utils.php                          16/16                      100.00%   ' . $eol;
-        $expected .= ' ------------------------------------------------------------------- -------------------------- ---------- ' . $eol;
-        $expected .= '  Overall Totals                                                      170/170                    100.00%   ' . $eol;
-        $expected .= ' ------------------------------------------------------------------- -------------------------- ----------';
+        $expected = '------------------------------------------------------------------------------- -------------------------- ---------- ' . $eol;
+        $expected .= '  File                                                                            Elements (Covered/Total)   Coverage  ' . $eol;
+        $expected .= ' ------------------------------------------------------------------------------- -------------------------- ---------- ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Application.php                                12/12                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\CoverageCheck.php                              82/82                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Utils.php                                      37/37                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Command\CoverageCheckCommand.php               79/79                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Exceptions\InvalidInputFileException.php       2/2                        100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Exceptions\NotAValidCloverFileException.php    2/2                        100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Exceptions\ThresholdOutOfBoundsException.php   2/2                        100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Style\CoverageCheckStyle.php                   14/14                      100.00%   ' . $eol;
+        $expected .= ' ------------------------------------------------------------------------------- -------------------------- ---------- ' . $eol;
+        $expected .= '  Overall Totals                                                                  230/230                    100.00%   ' . $eol;
+        $expected .= ' ------------------------------------------------------------------------------- -------------------------- ----------';
 
         self::assertSame($expected, trim($this->applicationTester->getDisplay()));
         self::assertSame(Command::SUCCESS, $this->applicationTester->getStatusCode());
@@ -391,17 +444,23 @@ final class CoverageCheckCommandTest extends TestCase
 
         $eol = PHP_EOL;
 
-        $expected = '------------------------------------------------------------------- -------------------------- ---------- ' . $eol;
-        $expected .= '  File                                                                Elements (Covered/Total)   Coverage  ' . $eol;
-        $expected .= ' ------------------------------------------------------------------- -------------------------- ---------- ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\Application.php                    10/10                      100.00%   ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\Command\CoverageCheckCommand.php   77/77                      100.00%   ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\CoverageCheck.php                  63/63                      100.00%   ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\Style\CoverageCheckStyle.php       4/4                        100.00%   ' . $eol;
-        $expected .= '  [...]\phpunit-coverage-check\src\Utils.php                          16/16                      100.00%   ' . $eol;
-        $expected .= ' ------------------------------------------------------------------- -------------------------- ---------- ' . $eol;
-        $expected .= '  Overall Totals                                                      170/170                    100.00%   ' . $eol;
-        $expected .= ' ------------------------------------------------------------------- -------------------------- ----------';
+        $expected = '------------------------------------------------------------------------ -------------------------- ---------- ' . $eol;
+        $expected .= '  File                                                                     Elements (Covered/Total)   Coverage  ' . $eol;
+        $expected .= ' ------------------------------------------------------------------------ -------------------------- ---------- ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Application.php                         12/12                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\CoverageCheck.php                       82/82                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Utils.php                               37/37                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Command\CoverageCheckCommand.php        79/79                      100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Exceptions\InvalidInputFileException.   2/2                        100.00%   ' . $eol;
+        $expected .= '  php                                                                                                           ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Exceptions\NotAValidCloverFileExcepti   2/2                        100.00%   ' . $eol;
+        $expected .= '  on.php                                                                                                        ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Exceptions\ThresholdOutOfBoundsExcept   2/2                        100.00%   ' . $eol;
+        $expected .= '  ion.php                                                                                                       ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Style\CoverageCheckStyle.php            14/14                      100.00%   ' . $eol;
+        $expected .= ' ------------------------------------------------------------------------ -------------------------- ---------- ' . $eol;
+        $expected .= '  Overall Totals                                                           230/230                    100.00%   ' . $eol;
+        $expected .= ' ------------------------------------------------------------------------ -------------------------- ----------';
 
         self::assertSame($expected, trim($this->applicationTester->getDisplay()));
         self::assertSame(Command::SUCCESS, $this->applicationTester->getStatusCode());